remove unused config options
