+++ 1.21/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v21.0.3-1.21] - 2024-08-07
+### Fixed
+- Fix anvils not breaking faster with the correct tool
+
 ## [v21.0.2-1.21] - 2024-07-07
 ### Changed
 - Updated for Puzzles Lib v21.0.8
