+++ 1.21/Common/src/main/java/fuzs/easyanvils/config/RenameAndRepairCost.java
@@ -1,0 +1,7 @@
+package fuzs.easyanvils.config;
+
+public enum RenameAndRepairCost {
+    VANILLA,
+    FIXED,
+    LIMITED
+}
