+++ 1.21/gradle.properties
@@ -5,7 +5,7 @@ copyBuildJar=true
 # Mod Attributes
 modId=easyanvils
 modName=Easy Anvils
-modVersion=21.0.2
+modVersion=21.0.3
 modAuthor=Fuzs
 modDescription=Be ready for overhauled anvils! Items stay, better name tags, many tweaks!
 modLicense=MPL-2.0
@@ -19,9 +19,9 @@ modForgeDisplayTest=MATCH_VERSION
 modFabricEnvironment=*
 
 # Version Catalog
-dependenciesVersionCatalog=1.21-v8
-dependenciesPuzzlesLibVersion=21.0.8
-dependenciesMinPuzzlesLibVersion=21.0.8
+dependenciesVersionCatalog=1.21-v22
+dependenciesPuzzlesLibVersion=21.0.19
+dependenciesMinPuzzlesLibVersion=21.0.19
 
 # Mod Publishing
 projectReleaseType=release
